feat: reorganize validation code into dedicated module (#302)
- Move validation logic from utilities.py to src/soso/validation/core.py
- Create output.py, logging.py, and cli.py for modular organization
- Update imports to use new validation module structure
- Prepare codebase for further validation enhancements
- Add tests/test_validation.py for validation core and output utilities
- Migrate existing validation tests
- Update documentation
