+++ src/soso/validation/cli.py
@@ -1,0 +1,1 @@
+"""CLI entry point"""
