+++ src/soso/validation/logging.py
@@ -1,0 +1,1 @@
+"""Logging setup"""
