+++ src/soso/validation/output.py
@@ -1,0 +1,1 @@
+"""Output formatting and writing."""
