+++ tests/test_utilities.py
@@ -1,15 +1,13 @@
-"""For testing the validator module."""
+"""For testing the utilities module."""
 
 import warnings
 from pathlib import Path
 from json import dumps
 import pytest
 from soso.utilities import (
-    validate,
     is_url,
     get_example_metadata_file_path,
     get_empty_metadata_file_path,
-    get_shacl_file_path,
     is_html,
     delete_null_values,
     delete_unused_vocabularies,
@@ -18,44 +16,6 @@ from soso.utilities import (
     as_numeric,
     guess_mime_type_with_fallback,
 )
-
-
-@pytest.mark.internet_required
-def test_validate_returns_warning_when_invalid(internet_connection):
-    """Test validate returns a warning when the graph is invalid."""
-    if not internet_connection:
-        pytest.skip("Internet connection is not available.")
-    with pytest.warns(UserWarning, match="Validation Report"):
-        validate("tests/incomplete.jsonld")
-
-
-@pytest.mark.internet_required
-def test_validate_returns_no_warning_when_valid(internet_connection):
-    """Test validate returns no warning when the graph is valid."""
-    if not internet_connection:
-        pytest.skip("Internet connection is not available.")
-    with warnings.catch_warnings(record=True) as list_of_warnings:
-        validate("tests/full.jsonld")
-        for warning in list_of_warnings:
-            assert not issubclass(warning.category, UserWarning)
-
-
-@pytest.mark.internet_required
-def test_validate_returns_true_when_valid(internet_connection):
-    """Test validate returns True when the graph is valid."""
-    if not internet_connection:
-        pytest.skip("Internet connection is not available.")
-    assert validate("tests/full.jsonld") is True
-
-
-@pytest.mark.internet_required
-def test_validate_returns_false_when_invalid(internet_connection):
-    """Test validate returns False when the graph is invalid."""
-    if not internet_connection:
-        pytest.skip("Internet connection is not available.")
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        assert validate("tests/incomplete.jsonld") is False
 
 
 def test_get_example_metadata_file_path_returns_path(strategy_names):
@@ -70,12 +30,6 @@ def test_get_empty_metadata_file_path_returns_path(strategy_names):
     for strategy in strategy_names:
         file_path = get_empty_metadata_file_path(strategy=strategy)
         assert isinstance(file_path, Path)
-
-
-def test_get_shacl_file_path_returns_path():
-    """Test that get_shacl_file_path returns a path."""
-    file_path = get_shacl_file_path()
-    assert isinstance(file_path, Path)
 
 
 def test_rm_null_values():
